Add files via upload
In this project i will  make an Online Quiz website or Quiz App with HTML, CSS and JavaScript. 

How this quiz app works:
In this quiz app we can add multiple questions and 4 answer choices for each questions. When you select any one answer the the answer button's background color will become green if it is correct answer and the background color will become red if it is wrong answer.

After selecting one answer you can not change the answer, you can only go for next question.

Display quiz score or result:
When user will submit the answer of last question and click on the next button. Then it will display the score.